update my real bomber malloc

+++ srcs/show_alloc_mem.c
@@ -12,7 +12,7 @@ static long long int 	show_tiny_mem()
 	while (block)
 	{
 		printf("%#X - %#X : %zu octets\n", (int) block->ptr, (int) block->ptr + (int) block->size, block->size);
-		total_size += block->size;
+		total_size += (long long int)block->size;
 		block = block->next;
 	}
 	return total_size;
@@ -30,7 +30,7 @@ static long long int	show_small_mem()
 	while ((block = find_next_allocated_block(block)))
 	{
 		printf("%#X - %#X : %zu octets\n", (int) block->ptr, (int) block->ptr + (int) block->size, block->size);
-		total_size += block->size;
+		total_size += (long long int)block->size;
 		block = block->next;
 	}
 	return total_size;
@@ -48,7 +48,7 @@ static long long int	show_large_mem()
 	while ((block = find_next_allocated_block(block)))
 	{
 		printf("%#X - %#X : %zu octets\n", (int) block->ptr, (int) block->ptr + (int) block->size, block->size);
-		total_size += block->size;
+		total_size += (long long int)block->size;
 		block = block->next;
 	}
 	return total_size;
@@ -64,5 +64,5 @@ void 				show_alloc_mem()
 	total_size += show_small_mem();
 	total_size += show_large_mem();
 
-	printf("Total : %zu\n", total_size);
+	printf("Total : %lld\n", total_size);
 }
